+++ testing/testing_zungqr.c
@@ -165,8 +165,8 @@ testing_zungqr_std( run_arg_list_t *args, int check )
     }
 
     /* Creates the matrices */
-    A    = malloc( LDA*K*sizeof(CHAMELEON_Complex64_t) );
-    Qlap = malloc( LDA*N*sizeof(CHAMELEON_Complex64_t) );
+    A    = malloc( (size_t) LDA*K*sizeof(CHAMELEON_Complex64_t) );
+    Qlap = malloc( (size_t) LDA*N*sizeof(CHAMELEON_Complex64_t) );
     CHAMELEON_Alloc_Workspace_zgels( M, K, &descT, P, Q );
 
     /* Fills the matrix with random values */
@@ -187,7 +187,7 @@ testing_zungqr_std( run_arg_list_t *args, int check )
 
     /* Checks the factorisation and orthogonality */
     if ( check ) {
-        CHAMELEON_Complex64_t *A0 = malloc( LDA*K*sizeof(CHAMELEON_Complex64_t) );
+        CHAMELEON_Complex64_t *A0 = malloc( (size_t) LDA*K*sizeof(CHAMELEON_Complex64_t) );
         CHAMELEON_zplrnt( M, K, A0, LDA, seedA );
 
         hres += check_zortho_std( args, M, N, Qlap, LDA );
